handling exception on delete

+++ src/main/java/com/nathanlucas/nscommerce/Services/ProductService.java
@@ -1,13 +1,17 @@
 package com.nathanlucas.nscommerce.Services;
 
+import com.nathanlucas.nscommerce.Services.exceptions.DatabaseException;
 import com.nathanlucas.nscommerce.Services.exceptions.ResourceNotFoundException;
 import com.nathanlucas.nscommerce.dtos.ProductDTO;
 import com.nathanlucas.nscommerce.entities.Product;
 import com.nathanlucas.nscommerce.repositories.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -40,15 +44,27 @@ public class ProductService {
 
     @Transactional
     public Product update(Long id, Product updatedProduct) {
-        Product entity = productRepository.getReferenceById(id);
-        modelMapper.map(updatedProduct, entity);
-        entity.setId(id);
-        return productRepository.save(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            modelMapper.map(updatedProduct, entity);
+            entity.setId(id);
+            return productRepository.save(entity);
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteById(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
 }
